Fix error - catch `ArrayIndexOutOfBoundsException` in parser instead of Command object
