+++ src/main/java/tfifteenfour/clipboard/logic/commands/editcommand/EditStudentCommand.java
@@ -1,26 +1,20 @@
 package tfifteenfour.clipboard.logic.commands.editcommand;
 
 import static java.util.Objects.requireNonNull;
-import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_COURSE;
 import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_NAME;
 import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_PHONE;
 import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_STUDENTID;
-import static tfifteenfour.clipboard.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.List;
-import java.util.Optional;
 
 import tfifteenfour.clipboard.commons.core.Messages;
 import tfifteenfour.clipboard.commons.core.index.Index;
-import tfifteenfour.clipboard.commons.util.CollectionUtil;
 import tfifteenfour.clipboard.logic.CurrentSelection;
 import tfifteenfour.clipboard.logic.PageType;
 import tfifteenfour.clipboard.logic.commands.CommandResult;
 import tfifteenfour.clipboard.logic.commands.exceptions.CommandException;
-import tfifteenfour.clipboard.logic.parser.ArgumentMultimap;
-import tfifteenfour.clipboard.logic.parser.ArgumentTokenizer;
-import tfifteenfour.clipboard.logic.parser.ParserUtil;
+import tfifteenfour.clipboard.logic.parser.NewEditCommandParser.EditStudentDescriptor;
 import tfifteenfour.clipboard.logic.parser.exceptions.ParseException;
 import tfifteenfour.clipboard.model.Model;
 import tfifteenfour.clipboard.model.course.Group;
@@ -39,7 +33,7 @@ import tfifteenfour.clipboard.model.task.Task;
 public class EditStudentCommand extends EditCommand {
     public static final String COMMAND_TYPE_WORD = "student";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Edits the details of the student identified "
+    public static final String MESSAGE_USAGE = COMMAND_WORD + COMMAND_TYPE_WORD + ": Edits the details of the student identified "
             + "by the index number used in the displayed student list. "
             + "Existing values will be overwritten by the input values.\n"
             + "Parameters: INDEX (must be a positive integer) "
@@ -47,66 +41,32 @@ public class EditStudentCommand extends EditCommand {
             + "[" + PREFIX_PHONE + "PHONE] "
             + "[" + PREFIX_EMAIL + "EMAIL] "
             + "[" + PREFIX_STUDENTID + "STUDENTID] "
-            + "[" + PREFIX_COURSE + "MODULE]..."
-            + "[" + PREFIX_TAG + "TAG]...\n"
             + "Example: " + COMMAND_WORD + " 1 "
             + PREFIX_PHONE + "91234567 "
             + PREFIX_EMAIL + "johndoe@example.com";
 
     public static final String MESSAGE_EDIT_PERSON_SUCCESS = "Edited Student: %1$s";
-    public static final String MESSAGE_NOT_EDITED = "At least one field to edit must be provided.";
     public static final String MESSAGE_DUPLICATE_PERSON = "This student already exists in the address book.";
 
 
 
     private final Index index;
-    private final String args;
-//    private final EditStudentDescriptor editStudentDescriptor;
-
-    private EditStudentDescriptor parseStudentInfo(String args) throws ParseException {
-        requireNonNull(args);
-        ArgumentMultimap argMultimap =
-                ArgumentTokenizer.tokenizePrefixes(args, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_STUDENTID,
-                        PREFIX_COURSE, PREFIX_TAG);
-
-        EditStudentDescriptor editStudentDescriptor = new EditStudentDescriptor();
-        if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
-            editStudentDescriptor.setName(ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get()));
-        }
-        if (argMultimap.getValue(PREFIX_PHONE).isPresent()) {
-            editStudentDescriptor.setPhone(ParserUtil.parsePhone(argMultimap.getValue(PREFIX_PHONE).get()));
-        }
-        if (argMultimap.getValue(PREFIX_EMAIL).isPresent()) {
-            editStudentDescriptor.setEmail(ParserUtil.parseEmail(argMultimap.getValue(PREFIX_EMAIL).get()));
-        }
-        if (argMultimap.getValue(PREFIX_STUDENTID).isPresent()) {
-            editStudentDescriptor.setStudentId(ParserUtil.parseStudentId(argMultimap.getValue(PREFIX_STUDENTID).get()));
-        }
-
-        if (!editStudentDescriptor.isAnyFieldEdited()) {
-            throw new ParseException(MESSAGE_NOT_EDITED);
-        }
-
-        return editStudentDescriptor;
-    }
+    private final EditStudentDescriptor editStudentDescriptor;
 
     /**
      * @param index of the student in the filtered student list to edit
      *
      */
     /**
+     *  @param index of the student in the filtered student list to edit
      *
-     * @param index of the student in the filtered student list to edit
-     * @param args Student information to be updated
      */
-    public EditStudentCommand(Index index, String args) {
+    public EditStudentCommand(Index index, EditStudentDescriptor editStudentDescriptor) {
         requireNonNull(index);
-        requireNonNull(args);
-//        requireNonNull(editStudentDescriptor);
+        requireNonNull(editStudentDescriptor);
 
         this.index = index;
-        this.args = args;
-//        this.editStudentDescriptor = new EditStudentDescriptor(editStudentDescriptor);
+        this.editStudentDescriptor = editStudentDescriptor;
     }
 
     @Override
@@ -126,7 +86,6 @@ public class EditStudentCommand extends EditCommand {
         }
 
         Student studentToEdit = lastShownList.get(index.getZeroBased());
-        EditStudentDescriptor editStudentDescriptor = parseStudentInfo(args);
         Student editedStudent = createEditedStudent(studentToEdit, editStudentDescriptor);
 
         if (!studentToEdit.isSameStudent(editedStudent) && model.hasStudent(editedStudent)) {
@@ -163,91 +122,6 @@ public class EditStudentCommand extends EditCommand {
                 remark);
     }
 
-    /**
-     * Stores the details to edit the student with. Each non-empty field value will replace the
-     * corresponding field value of the student.
-     */
-    public static class EditStudentDescriptor {
-        private Name name;
-        private Phone phone;
-        private Email email;
-        private StudentId studentId;
-
-        public EditStudentDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditStudentDescriptor(EditStudentDescriptor toCopy) {
-            setName(toCopy.name);
-            setPhone(toCopy.phone);
-            setEmail(toCopy.email);
-            setStudentId(toCopy.studentId);
-        }
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(name, phone, email, studentId);
-        }
-
-        public void setName(Name name) {
-            this.name = name;
-        }
-
-        public Optional<Name> getName() {
-            return Optional.ofNullable(name);
-        }
-
-        public void setPhone(Phone phone) {
-            this.phone = phone;
-        }
-
-        public Optional<Phone> getPhone() {
-            return Optional.ofNullable(phone);
-        }
-
-        public void setEmail(Email email) {
-            this.email = email;
-        }
-
-        public Optional<Email> getEmail() {
-            return Optional.ofNullable(email);
-        }
-
-        public void setStudentId(StudentId studentId) {
-            this.studentId = studentId;
-        }
-
-        public Optional<StudentId> getStudentId() {
-            return Optional.ofNullable(studentId);
-        }
-
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditStudentDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditStudentDescriptor e = (EditStudentDescriptor) other;
-
-            return getName().equals(e.getName())
-                    && getPhone().equals(e.getPhone())
-                    && getEmail().equals(e.getEmail())
-                    && getStudentId().equals(e.getStudentId());
-        }
-    }
-
     @Override
     public boolean equals(Object other) {
         // short circuit if same object
@@ -263,7 +137,7 @@ public class EditStudentCommand extends EditCommand {
         // state check
         EditStudentCommand e = (EditStudentCommand) other;
         return index.equals(e.index)
-                && args.equals(e.args);
+                && editStudentDescriptor.equals(e.editStudentDescriptor);
     }
 
 }
